+++ tests/Support/NamedEnum.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Tests\Support;
+
+enum NamedEnum
+{
+    case one;
+    case two;
+    case three;
+}